Revert "test moving up a level"
This reverts commit aef3c5b75becf6aaac6ede39f1881bea7a07b4ac.
